chore: refactor module config
