+++ src/config/mod.rs
@@ -1,2 +1,2 @@
-pub mod entity;
-pub mod read;
+pub mod helper;
+pub mod schema;
